feat(k3): save BL31 before to enter sleep
Before to enter sleep BL31 is saved if a K3_LPM_DDR_SAVE_ADDRESS address
has been set.
BL31 need to be saved if it runs in SRAM, but not if it runs in DDR.

No need to explicitly flush the caches, BL31 will ask sysfw to flush L2
cache once the last core of cluster will be powered down.

Change-Id: If26a91936c0bb52563697b66904a19ef522922f9
Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>
Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>