feat(k3): Add TIFS encryption message
At suspend, BL31 with its context will be encrypted by TIFS in DDR.
Encryption is needed for security matters, so that the BL31 is not
modified before entering suspend or early at resume.

We only need the encryption function here because the decryption message
will be send by the R5 SPL at resume.

Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>